+++ src/main/java/racingcar/controller/MoveStrategy.java
@@ -1,4 +1,4 @@
-package racingcar.service;
+package racingcar.controller;
 
 @FunctionalInterface
 public interface MoveStrategy {
