+++ src/main/java/racingcar/controller/RandomMove.java
@@ -1,4 +1,4 @@
-package racingcar.service;
+package racingcar.controller;
 
 import java.util.Random;
